+++ src/main/java/com/multi/mini/qna/controller/QnaController.java
@@ -57,21 +57,31 @@ public class QnaController {
 
         /* 파일명 변경 처리 */
         String originFileName = singleFile.getOriginalFilename();
-        String ext = originFileName.substring(originFileName.lastIndexOf("."));
-        String savedName = UUID.randomUUID().toString().replace("-", "") + ext;
+        if(originFileName != null && !originFileName.isEmpty()){
+            String ext = originFileName.substring(originFileName.lastIndexOf("."));
+            String savedName = UUID.randomUUID().toString().replace("-", "") + ext;
 
-        /* 파일을 저장한다. */
-        try {
-            singleFile.transferTo(new File(filePath + "\\" + savedName));
-            model.addAttribute("savedName", savedName);
-            qnaDTO.setImg(savedName);
-            qnaService.insertQna(qnaDTO);
-        } catch (Exception e) {
-            System.out.println("qna insert error : " + e);
-            /* 실패시 파일 삭제 */
-            new File(filePath + "\\" + savedName).delete();
-            model.addAttribute("message", "파일 업로드 실패!!");
+            /* 파일을 저장한다. */
+            try {
+                singleFile.transferTo(new File(filePath + "\\" + savedName));
+                model.addAttribute("savedName", savedName);
+                qnaDTO.setImg(savedName);
+                qnaService.insertQna(qnaDTO);
+            } catch (Exception e) {
+                System.out.println("qna insert error : " + e);
+                /* 실패시 파일 삭제 */
+                new File(filePath + "\\" + savedName).delete();
+                model.addAttribute("message", "파일 업로드 실패!!");
+            }
         }
+        else{
+            try {
+                qnaService.insertQna(qnaDTO);
+            } catch (Exception e) {
+                System.out.println("qna insert error : " + e);
+            }
+        }
+
 
         // 서비스 계층에 게시글 저장 요청
         return "redirect:/home";
